requirements.txt
pillow
transformers
torch
accelerate
torchvision
PyMuPDF
reportlab
